splits TPFRequestProcessorImpl into TPFRequestProcessorBase and TPFRequestProcessorImpl
